navbar works in main page
