+++ client/src/App.css
@@ -15,16 +15,6 @@
   font-size: large;
 }
 
-a, .navbar-brand, .navbar-nav .nav-link {
-  color: #bbb;
-}
-
-.navbar {
- 
-  margin-top: auto;
-  background-color: #222;
-  
-}
 #top-filler {
   margin-bottom: auto;
 }
