+++ client/src/components/Nav/style.css
@@ -1,3 +1,12 @@
 .Navigation {
    padding-bottom: 300px;
-  }
+}
+
+a, .navbar-brand, .navbar-nav .nav-link {
+   color: #bbb;
+}
+ 
+.navbar {
+   margin-top: auto;
+   background-color: #222;   
+}
